DEV-5064: Rename command
